Moved tabs.js into the UndergraduateView folder on navigation
